style: retire `type.py`
